requirements.txt
# Environment: rec_face
# Python: 3.10.19
#
# Non-Pip Dependencies (Conda recommended):
#   conda install -c conda-forge cuda-version=11.8 cudatoolkit=11.8.0 cudnn=8.9.7.29
#
# pip instructions:
#   pip install -r requirements.txt
#
# Index for CUDA 11.8 wheels:
--extra-index-url https://download.pytorch.org/whl/cu118

annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
dnspython==2.8.0
exceptiongroup==1.3.1
faiss-cpu==1.13.1
fastapi==0.128.0
filelock==3.20.0
Flask==3.0.3
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
kafka==1.3.5
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mkl==2021.4.0
mpmath==1.3.0
networkx==2.8.8
numpy==1.26.4
onnxruntime-gpu==1.16.3
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.0.3
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.10.1
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
safetensors==0.7.0
scikit-image==0.24.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
tbb==2021.13.1
tifffile==2025.5.10
timm==1.0.22
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.235
ultralytics-thop==2.0.18
urllib3==2.6.1
uvicorn==0.40.0
websockets==16.0
Werkzeug==3.1.4
yacs==0.1.8
